+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 - Breaking change: removed `all`, `test` and `typescript-react` configs. Just importing the index config
   will enable all configs. For just the specific configs import `@bonniernews/eslint-config/js`,
   `@bonniernews/eslint-config/ts`, `@bonniernews/eslint-config/jsx`, `@bonniernews/eslint-config/test-js`,
-  `@bonniernews/eslint-config/test-ts`, or `@bonniernews/eslint-config/tsx`.
+  `@bonniernews/eslint-config/test-ts`, or `@bonniernews/eslint-config/tsx`. When creating your own set of
+  rules different globals (as an example `globals.browser`) is available from `@bonniernews/eslint-config/globals`.
 
 ## 1.2.0